% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculFastSpectral_p.r
\name{computeGap}
\alias{computeGap}
\title{Compute gap between eigenvalues of a similarity matrix}
\usage{
computeGap(similarity, Gmax)
}
\arguments{
\item{similarity}{a similarity matrix.}

\item{Gmax}{the maximum gap value allowed (only the first Gmax eigenvalues will be taken into account).}
}
\value{
The function returns a list containing the following components:

\item{gap}{a vector indicating the gap between similarity matrix eigenvalues (the gap between the 2 first eigenvalues is set to 0)}

\item{Kmax}{an integer indicating the index of the highest gap (the highest gap is between the Kmax-th and the (Kmax+1)-th eigenvalues)}
}
\description{
Compute the highest gap between eigenvalues of a similarity matrix.
The 2 first eigenvalues are considered as equal to each other (the gap between the 2 first eigenvalues is set to 0).
}
\examples{
similarity<-matrix(runif(100),ncol=10)
Gap<-computeGap(matrix(runif(100),ncol=10),10)
plot(Gap$gap,type="h",main=paste("Gap =",Gap$K))

}
\author{
Kevin Rousseeuw ? Emilie Poisson-Caillault ?
}

