% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculFastSpectral_p.r
\name{cutCalculation}
\alias{cutCalculation}
\title{cutCalculation function}
\usage{
cutCalculation(similarity, label, K)
}
\arguments{
\item{similarity}{a similarity matrix.}

\item{label}{vector of cluster sequencing. (prototypes ???)}

\item{K}{number of clusters ???.}
}
\value{
The function returns a list containing:
\item{mncut}{the inter cluster cut, i.e. K - sum(ratioCutVol).}
\item{ratioCutVol}{vector of intra cluster cuts, one component per cluster.}
}
\description{
I don't know.
}
\details{
intra cluster cut :
\deqn{Cut(g_{k},g_{l}) = \sum_{i=1,x(i)\in g_{k}}^{N_{p}}\sum_{j=1,x(j)\in g_{l}}^{N_{p}}w(x(i),x(j)) }
}
\author{
Kevin Rousseeuw ? Emilie Poisson-Caillault ?
}

