% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateProcessing.r
\name{dateProcessing}
\alias{dateProcessing}
\title{Create a vector of observation times}
\usage{
dateProcessing(data)
}
\arguments{
\item{data}{dataframe with a column "Dates" in "YYYY-MM-DD" format and a column "Hours" in "HH:MM:SS" format.}
}
\value{
numeric vector (use \code{\link[chron]{chron}} function to get the vector in date format).
}
\description{
Paste date and time of observations in a vector.
}

