% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMMparams.r
\name{emissionMatrix}
\alias{emissionMatrix}
\title{Emission matrix estimation}
\usage{
emissionMatrix(states, symbols)
}
\arguments{
\item{states}{a numeric vector of state sequencing.}

\item{symbols}{a numeric vector of symbol sequencing.}
}
\value{
Estimated emission matrix.
}
\description{
This function estimates the emission matrix of a Hidden Markov Model from vectors of state and symbol sequencing.
}
\examples{
states<-c(1,1,3,2,1,2,1,3)
symbols<-c(4,1,3,1,4,4,4,2)
B<-emissionMatrix(states,symbols)
B
}
\seealso{
\code{\link{HMMparams}}
}

