% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMMparams.r
\name{HMMparams}
\alias{HMMparams}
\title{Hidden Markov Model parameter estimation}
\usage{
HMMparams(stateSeq, symbolSeq)
}
\arguments{
\item{stateSeq}{a numeric vector of state sequencing.}

\item{symbolSeq}{a numeric vector of symbol sequencing.}
}
\value{
HMMparams returns a list containing :
\item{trans}{The transition matrix.}
\item{emis}{The emission matrix.}
\item{startProb}{The vector of initial probability distribution (initial states are supposed equiprobable).}
}
\description{
This function is used by the \code{\link{uHMMinterface}} to estimate parameters of a Hidden Markov Model.
}
\seealso{
\code{\link{transitionMatrix}} \code{\link{emissionMatrix}}
}

