% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculFastSpectral_p.r
\name{KpartitionNg}
\alias{KpartitionNg}
\title{KpartitionNg function}
\source{
Ng Andrew, Y., M. I. Jordan, and Y. Weiss. "On spectral clustering: analysis and an algorithm [C]." Advances in Neural Information Processing Systems (2001).
}
\usage{
KpartitionNg(similarity, K)
}
\arguments{
\item{similarity}{matrix of similarity.}

\item{K}{number of clusters.}
}
\value{
The function returns a list containing:
\item{label}{vector of cluster sequencing.}
\item{centres}{matrix of cluster centers in the space of the K first normalised eigen vectors.}
\item{vecteursPropresProjK}{matrix containing, in columns, the K first normalised eigen vectors of the similarity matrix.}
\item{valeursPropresK}{vector containing the K first eigen values of the similarity matrix.}
\item{vecteursPropres}{matrix containing, in columns, eigen vectors of the similarity matrix.}
\item{valeursPropres}{vector containing eigen values of the similarity matrix.}
\item{inertieZ}{vector of within-cluster sum of squares, one component per cluster.}
}
\description{
Perform spectral classification on the similarity matrix of a dataset (Ng et al. (2001) algorithm), using kmeans algorithm on data projected in the space of its K first eigen vectors.
}
\author{
Kevin Rousseeuw ? Emilie Poisson Caillault ?
}

