% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kppv.r
\name{kppv}
\alias{kppv}
\title{k-Nearest Neighbours}
\usage{
kppv(train, test = NULL, label = NULL, K = 1)
}
\arguments{
\item{train}{matrix or data frame of training set.}

\item{test}{matrix or data frame of test set. If NULL, the training set is also used as test set.}

\item{label}{data labels.}

\item{K}{number of neighbours considered.}
}
\value{
The function returns a list with the following components:
\item{D}{matrix of squared root of the distances between observations and their nearest neighbours.}
\item{idx}{Index of K nearest neighbours of each observation.}
\item{classes}{? vector giving the classe of each observation.}
}
\description{
This function performs the k-Nearest Neighbour algorithm without class estimation, but only computation of distances and neighbours.
}
\author{
Emilie Poisson-Caillault ? Kevin Rousseeuw ?
}

