% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctionKppvObsCentre.r
\name{kppvObsCentre}
\alias{kppvObsCentre}
\title{Affect nearest symbols of a data set}
\usage{
kppvObsCentre(data, gravityCenters, normParams = NULL)
}
\arguments{
\item{data}{dataframe of cases for which nearest symbols must be affected.}

\item{gravityCenters}{matrix of symbol gravity centers.}

\item{normParams}{a list containing a vector of means and a vector of standard deviations of columns of the training dataset (only if classification has been performed on normalized dataset).}
}
\value{
The vector of affected symbols (the first symbol is the one which is affected to observations with NA's).
}
\description{
This function is used by the \code{\link{uHMMinterface}} to affect symbols to data, using KNN algorithm, after the classification step.
}
\seealso{
\code{\link[class]{knn}}
}

