% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marelData.r
\name{MarelCarnot}
\alias{MarelCarnot}
\title{MarelCarnot dataset}
\format{A data frame with 131487 rows and 16 columns.}
\source{
Lefebvre Alain (2015). MAREL Carnot data and metadata from Coriolis Data Centre. SEANOE. \url{http://doi.org/10.17882/39754}
}
\usage{
MarelCarnot
}
\description{
The MarelCarnot data set gives the measurements of 14 physico-chemical and biological parameters performed by 
the Marel-Carnot station (Boulogne-sur-Mer, France), at high frequency resolution.
}
\details{
\tabular{lll}{
Dates \tab date of measurement \tab (YYYY:MM:DD) \cr
Hours \tab time of measurement \tab (HH:MM:SS) \cr
C_NI1 \tab nitrate concentration \tab (in \eqn{\mu}mol/L) \cr
C_PO1 \tab phosphate concentration \tab (in \eqn{\mu}mol/L) \cr
C_O21 \tab corrected dissolved oxygen \tab (in mg/L) \cr
C_SI1 \tab silicate concentration \tab (in \eqn{\mu}mol/L) \cr
CSAL1 \tab salinity \tab (in PSU) \cr
CSAT1 \tab oxygen saturation \tab (in \%)  \cr
ETCO1 \tab air temperature \tab (in degrees Celsius) \cr
E_LU1 \tab P.A.R \tab (in \eqn{\mu}mol of photons/s/m2) \cr
E_O21 \tab uncorrected dissolved oxygen \tab (in mg/L) \cr
E_PH1 \tab pH \cr
E_TU1 \tab turbidity \tab (in NTU) \cr
ECHL1 \tab fluorescence \tab (in FFU) \cr
E__TA \tab water temperature \tab (in degrees Celsius) \cr
XMAHH \tab water level \tab (in m) \cr
}
}

