% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctionMarkovViterbiEstimation.r
\name{MarkovViterbiEstimation}
\alias{MarkovViterbiEstimation}
\title{Compute the most probable path of states}
\usage{
MarkovViterbiEstimation(nbStates, nbSymbols, transProbs, emissionProbs,
  startProb, symbolSeq)
}
\arguments{
\item{nbStates}{number of states (not including the NA state).}

\item{nbSymbols}{number of symbols (not including the NA symbol).}

\item{transProbs}{matrix containing the transition probabilities between the states.}

\item{emissionProbs}{matrix containing the emission probabilities of the states.}

\item{startProb}{vector of initial probability distribution of the states.}

\item{symbolSeq}{vector of symbol sequence.}
}
\value{
MarkovViterbiEstimation returns a list containing:
\item{estimatedHMM}{A HMM, which is a list of 5 elements (see \code{\link[HMM]{initHMM}}).}
\item{stateSeq}{The estimated state sequence.}
}
\description{
This function is used by the MMCNS interface to perfom the \code{\link[HMM]{viterbi}} algorithm,
 which computes the most probable path of states for a sequence of symbols and a given Hidden Markov Model.
}
\seealso{
\code{\link[HMM]{viterbi}} \code{\link[HMM]{initHMM}}
}

