% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMMparams.r
\name{transitionMatrix}
\alias{transitionMatrix}
\title{Transition matrix estimation}
\usage{
transitionMatrix(states)
}
\arguments{
\item{states}{a numeric vector of state sequencing.}
}
\value{
Estimated transition matrix.
}
\description{
This function estimates the transition matrix of a (Hidden) Markov Model from a vector of state sequencing.
}
\examples{
states<-c(1,1,3,2,1,2,1,3)
A<-transitionMatrix(states)
A
}
\seealso{
\code{\link{HMMparams}}
}

