\name{uHMM-package}
\alias{uHMM-package}
\docType{package}
\title{
Construct an unsupervised Hidden Markov Model
}
\description{
This package proposes an interface to detect usual or extreme events in a dataset and to characterize their dynamic, by building an unsupervised Hidden Markov Model (use \code{\link{uHMMinterface}} to launch the interface).
Functions can also be used out of the interface to build an uHMM.
}
\details{

\tabular{ll}{
Package: \tab uHMM \cr
Version: \tab 1.0 \cr
Date: \tab 2016-04-13 \cr
Depends: \tab  R (>= 3.0.0), stats, grDevices \cr
Import: \tab  tcltk, tcltk2, tkrplot, HMM, clValid, class, cluster, FactoMineR, corrplot, chron \cr
License: \tab GPL (>=2) \cr
LazyLoad: \tab yes\cr

}

}
\author{
Emilie Poisson-Caillault and Paul Ternynck 

Maintainer: <emilie.caillault@lisic.univ-littoral.fr>
}
\source{
Rousseeuw, Kevin, et al. "Hybrid hidden Markov model for marine environment monitoring." Selected Topics in Applied Earth Observations and Remote Sensing, IEEE Journal of 8.1 (2015): 204-213.

%Zelnik-Manor, Lihi, and Pietro Perona. "Self-tuning spectral clustering." Advances in neural information processing systems. 2004.


}