% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1uHMMinterface.r
\name{uHMMinterface}
\alias{uHMMinterface}
\title{Graphical Interface to Build an uHMM}
\usage{
uHMMinterface(uHMMenv = NULL)
}
\arguments{
\item{uHMMenv}{an environment in which data and results will be stored. If NULL, a local environment will be created.}
}
\value{
Results are saved in the directory chosen by the user.
}
\description{
A user-friendly interface to detect usual or extreme events in a dataset and to characterize their dynamic,
 by building an unsupervised Hidden Markov Model.
}
\references{
Rousseeuw, Kevin, et al. "Hybrid hidden Markov model for marine environment monitoring." Selected Topics in Applied Earth Observations and Remote Sensing, IEEE Journal of 8.1 (2015): 204-213.
}

